requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
torch>=2.2.0
xgboost>=2.0
lightgbm>=4.3
shap>=0.44
lime>=0.2.0.1
librosa>=0.10
opensmile>=2.5.0
antropy>=0.1.6
groq>=0.9.0
imbalanced-learn>=0.12
optuna>=3.5
joblib>=1.3
plotly>=5.18
seaborn>=0.13
matplotlib>=3.8
statsmodels>=0.14
pywavelets>=1.6
dcurves>=0.4
tqdm>=4.66
tab-transformer-pytorch
pyyaml>=6.0
